% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angstroms-package.R
\docType{package}
\name{angstroms-package}
\alias{angstroms-package}
\title{angstroms: Tools for 'ROMS' the Regional Ocean Modeling System}
\description{
Helper functions for working with Regional Ocean Modeling System 'ROMS' output. See \url{https://www.myroms.org/} for more information about 'ROMS'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AustralianAntarcticDivision/angstroms/}
  \item Report bugs at \url{https://github.com/AustralianAntarcticDivision/angstroms/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

Authors:
\itemize{
  \item Ben Raymond
  \item Hernan G. Arango (wrote original Matlab implementation of romsdepth)
  \item Ryan Morse (wrote original R version of create_transport)
}

}
\keyword{internal}
