% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angstroms-package.R
\docType{data}
\name{antarctica}
\alias{antarctica}
\title{Antarctica simple coastline.}
\description{
Taken from "rnaturalearth::countries110". Loaded as a live \link[terra:SpatVector-class]{terra::SpatVector}
via active binding from a bundled GeoPackage.
}
