% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/romsmap.R
\name{coords_points}
\alias{coords_points}
\title{Create SpatialPoints from coordinate arrays (deprecated)}
\usage{
coords_points(x, ...)
}
\arguments{
\item{x}{two layer SpatRaster with longitude and latitude values}

\item{...}{ignored}
}
\value{
matrix of coordinates with columns x, y
}
\description{
Use \code{coords_points} to get a matrix of coordinates from a two-layer coord SpatRaster. Returns a
two column matrix with columns x (longitude) and y (latitude).
}
\examples{
pts <- coords_points(ice_coords)
head(pts)
}
