% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupling.R
\name{coupling-deprecated}
\alias{coupling-deprecated}
\title{Create SpatialPoints (deprecated).}
\description{
This function is kept for backward compatibility. See \code{\link[=coords_points]{coords_points()}} which now returns
a plain matrix. For the old SpatialPoints return, convert manually.
}
