% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angstroms-package.R
\docType{data}
\name{ice_fake}
\alias{ice_fake}
\alias{ice_coords}
\title{Fake model data.}
\description{
\code{ice_coords} and \code{ice_fake} are generated from a projected map of southern Ocean sea ice data.
These are loaded as live \link[terra:SpatRaster-class]{terra::SpatRaster} objects via active bindings (not serialized .rda).
}
\details{
The coords layer is the longitude and latitude values for the centres of the polar cells. This is very loosely
analogous to the coordinate arrays used by ROMS data, included here for working examples, illustration and code tests.

The proper metadata for these layers is
"-3950000, 3950000, -3950000, 4350000  (xmin, xmax, ymin, ymax)"

"+proj=stere +lat_0=-90 +lat_ts=-70 +lon_0=0 +k=1 +x_0=0 +y_0=0 +a=6378273 +b=6356889.449 +units=m +no_defs"
}
