% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc-utils.r
\name{ncraster}
\alias{ncraster}
\title{Read an arbitrary 2D or 3D slice from NetCDF as a RasterBrick}
\usage{
ncraster(x, varname, slice)
}
\arguments{
\item{x}{ROMS file name}

\item{varname}{variable name}

\item{slice}{index, specified with NA for the index to read all steps}
}
\description{
Read an arbitrary 2D or 3D slice from NetCDF as a RasterBrick
}
