% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c-grid.r
\name{plot_cgrid}
\alias{plot_cgrid}
\title{https://www.myroms.org/wiki/easygrid
The rho-grid represents the grid centers (red dots)
The u-grid represents the grid East-West sides (blue triangles)
The v-grid represents the grid North-South sides (green triangles)
The psi-grid represents the grid corners (purple crosses)
A useable Thredds example is here
https://ecowatch.ncddc.noaa.gov/thredds/dodsC/northwest_fisheries/roms/nep_c.nc}
\usage{
plot_cgrid(x, ex = extent(0, 15, 0, 20), include = c("u", "v", "rho",
  "psi"), cell = TRUE, ...)
}
\arguments{
\item{x}{ROMS file}

\item{ex}{extent in index space}

\item{include}{grid elements to plot (all by default, uv, rho, psi)}

\item{cell}{draw the cells (defaults to TRUE)}

\item{...}{arguments passed to plot}
}
\description{
https://www.myroms.org/wiki/easygrid
The rho-grid represents the grid centers (red dots)
The u-grid represents the grid East-West sides (blue triangles)
The v-grid represents the grid North-South sides (green triangles)
The psi-grid represents the grid corners (purple crosses)
A useable Thredds example is here
https://ecowatch.ncddc.noaa.gov/thredds/dodsC/northwest_fisheries/roms/nep_c.nc
}
