% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupling2.R
\name{project_to}
\alias{project_to}
\title{Convenience function to transform map projection .}
\usage{
project_to(x, to)
}
\arguments{
\item{x}{object to transform}

\item{to}{object with a map projection}
}
\value{
`x`, transformed
}
\description{
Transform `x` to whatever the projection of `to` is.
}
