% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/romsmap.R
\name{romsboundary}
\alias{romsboundary}
\alias{databoundary}
\title{Boundary polygon from raster of coordinates.}
\usage{
romsboundary(cds)

databoundary(x, mask = NULL, ...)
}
\arguments{
\item{cds}{two-layer SpatRaster}

\item{x}{a thing}

\item{mask}{logical}

\item{...}{dots}
}
\description{
Create a boundary polygon by tracking around coordinates stored in a SpatRaster.
}
\details{
The first layer is treated as the X coordinate, second as Y.
}
\examples{
ice_grid_boundary <- romsboundary(ice_coords)
## boundary is a wk_polygon in lon/lat space
ice_grid_boundary
}
