% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc-utils.r
\name{romscoords_z}
\alias{romscoords_z}
\title{convert the depth ramp Cs_r, h (bottom depth), and cell number
to a correctly oriented layer of depth values}
\usage{
romscoords_z(x, cell)
}
\description{
convert the depth ramp Cs_r, h (bottom depth), and cell number
to a correctly oriented layer of depth values
}
