% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so_files.r
\name{so_files}
\alias{so_files}
\title{Find and return data base of files}
\usage{
so_files(files = NULL)
}
\arguments{
\item{files}{the files}
}
\value{
data frame of files
}
\description{
Find and return data base of files
}

