% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupling2.R
\name{coords_points}
\alias{coords_points}
\title{Create SpatialPoints.}
\usage{
coords_points(x, ...)
}
\arguments{
\item{x}{two layer \code{RasterBrick} with longitude and latitude values}

\item{...}{ignored}
}
\value{
\code{SpatialPoints}
}
\description{
Convenience wrapper around \code{SpatialPoints} for a two layer brick with longitude and latitude coordinate arrays.
}
\examples{
## library(raadtools)
##coords_points(romscoords(cpolarfiles()$fullname[1]))

pts <- coords_points(ice_coords)
}
