% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc-utils.R
\name{rawdata}
\alias{rawdata}
\alias{rawdata.character}
\alias{rawdata.NetCDF}
\title{Read the variable as is}
\usage{
rawdata(x, varname, ..., native = FALSE)

\method{rawdata}{character}(x, varname, ..., native = FALSE)

\method{rawdata}{NetCDF}(x, varname, ..., native = FALSE)
}
\arguments{
\item{x}{netcdf file path}

\item{varname}{variable name}

\item{...}{dots (ignored)}

\item{native}{maintain values in internal storge mode (default is \code{FALSE})}
}
\description{
Read the variable as is
}
