% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angstroms.r
\name{romsboundary}
\alias{romsboundary}
\title{Create a boundary polygon by tracking around coordinates stored in a RasterStack
 
The first layer is treated as the X coordinate, second as Y.}
\usage{
romsboundary(cds)
}
\arguments{
\item{cds}{two-layer Raster}
}
\description{
Create a boundary polygon by tracking around coordinates stored in a RasterStack
 
The first layer is treated as the X coordinate, second as Y.
}
