% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oms.r
\name{romscoords}
\alias{romscoords}
\title{Extract coordinate arrays from ROMS.}
\usage{
romscoords(x, spatial = c("lon_u", "lat_u"))
}
\arguments{
\item{x}{ROMS file name}

\item{spatial}{names of coordinate variables (e.g. lon_u, lat_u)}
}
\value{
\code{\link[raster]{RasterStack}}
}
\description{
Returns a RasterStack of the given variable names.
}
\examples{
\dontrun{
  coord <- romscoord("roms.nc")
}
}

