% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angstroms.r
\name{romscoords}
\alias{romscoords}
\title{Extract coordinate arrays from ROMS.}
\usage{
romscoords(x, spatial = c("lon_u", "lat_u"), ncdf = TRUE,
  transpose = FALSE, ...)
}
\arguments{
\item{x}{ROMS file name}

\item{spatial}{names of coordinate variables (e.g. lon_u, lat_u)}

\item{ncdf}{default to NetCDF no matter what file name}

\item{transpose}{the extents (ROMS is FALSE, Access is TRUE)}

\item{...}{unused}
}
\value{
RasterStack with two layers of the 2D-variables
}
\description{
Returns a RasterStack of the given variable names.
}
\examples{
\dontrun{
  coord <- romscoord("roms.nc")
}
}

