% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oms.r
\name{romsdata}
\alias{romsdata}
\title{Extract a data lyaer from ROMS by name and slice.}
\usage{
romsdata(x, varname, slice = c(1, 1))
}
\arguments{
\item{x}{ROMS file name}

\item{varname}{name of ROMS variable}

\item{slice}{index in w and t (depth and time), defaults to first encountered}
}
\value{
\code{\link[raster]{RasterLayer}}
}
\description{
Extract a data lyaer from ROMS by name and slice.
}

