% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc-utils.r
\name{romsdata}
\alias{romsdata}
\title{Extract a data layer from ROMS by name and slice.}
\usage{
romsdata(x, varname, slice = c(1, 1), ncdf = TRUE, ...)
}
\arguments{
\item{x}{ROMS file name}

\item{varname}{name of ROMS variable}

\item{slice}{index in w and t (depth and time), defaults to first encountered}

\item{ncdf}{default to \code{TRUE}, set to \code{FALSE} to allow raster format detection \code{\link[raster]{brick}}}
}
\value{
\code{\link[raster]{RasterLayer}}
}
\description{
Maybe this replaced by rastergetslice??
Returns a single slice 2D layer
}

