% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angstroms.r
\name{romshcoords}
\alias{romshcoords}
\title{Coordinates at depth
\code{S} and \code{h} are the  names of the appropriate variables}
\usage{
romshcoords(x, S = "Cs_r", depth = "h")
}
\arguments{
\item{x}{ROMS file name}

\item{S}{of S-coordinate stretching curve at RHO-points}

\item{h}{bathymetry at RHO-points}
}
\value{
\code{\link[raster]{RasterStack}} with a layer for every depth
}
\description{
Coordinates at depth
\code{S} and \code{h} are the  names of the appropriate variables
}

