% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angstroms.r
\name{romshcoords}
\alias{romshcoords}
\title{Coordinates at depth}
\usage{
romshcoords(x, S = "Cs_r", depth = "h")
}
\arguments{
\item{x}{ROMS file name}

\item{S}{of S-coordinate stretching curve at RHO-points}

\item{depth}{depth thing}
}
\value{
RasterStack with a layer for every depth
}
\description{
\code{S} and \code{h} are the  names of the appropriate variables
}
