% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{romsrotate}
\alias{romsrotate}
\alias{romsangle}
\title{Rotate 'ROMS' vectors}
\usage{
romsrotate(uv, angle, ...)

romsangle(x, varname = "angle", ...)
}
\arguments{
\item{uv}{raster brick of two layers with 'u' ("east-west") and 'v' ("north-south") vector components}

\item{angle}{raster layer of the 'angle' variable from 'ROMS'}

\item{...}{ignored currently}

\item{x}{filename}

\item{varname}{variable name from file 'x'}
}
\description{
\code{romsrotate} performs the rotation on a 2-layer brick
\code{romsangle} reads the (likely) angle variable
}
\references{
\href{https://www.myroms.org/forum/viewtopic.php?f=3&t=295}{ROMS website}
}
