% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{romsrotate}
\alias{romsrotate}
\alias{romsangle}
\title{Rotate roms vectors}
\usage{
romsrotate(uv, angle)

romsangle(x, varname = "angle", ...)
}
\description{
\code{romsrotate} performs the rotation on a 2-layer brick
\code{romsangle} reads the (likely) angle variable
}
\references{
\href{https://www.myroms.org/forum/viewtopic.php?f=3&t=295}{ROMS website}
}
